Adds Ansible output parser
Ansible output with -j option gives a sequence of json "events" to
stdout. This patch adds classes for AnsibleEvents and AnsibleEventList
to hold all of the events outputted by an ansible-runner execution.

In a future patch we will read the status on the last event to set
pass/fail on the command. Also, we can determine the event which failed
and log it.

Signed-off-by: Tim Rozet <tdrozet@gmail.com>